many changes & add unit testing

+++ src/swagger.ts
@@ -2,27 +2,19 @@
 import { Application } from "npm:express@4.18.2";
 import SwagInit from "npm:swagger-autogen@2.23.1";
 
-import { Server, Swagger } from "./env.ts";
-import { modules_dir, swagger_html_file, swagger_json_file } from "./path.ts";
+import { Swagger } from "./env.ts";
+import { modules_dir, swagger_json_file, app_file } from "./path.ts";
 
-// import SwaggerUI from "npm:swagger-ui@5.7.2";
-// const swagger = SwaggerUI({
-//   url: "/swagger.json",
-//   dom_id: "#swagger-ui",
-// });
-// console.log({ swagger });
+import { StatusCodes } from "npm:http-status-codes@2.2.0";
 
-// import { SwaggerUIBundle } from "npm:swagger-ui-dist@5.7.2";
-// var SwaggerUIBundle = require('swagger-ui-dist').SwaggerUIBundle
-// const ui = SwaggerUIBundle({
-//   url: "/swagger.json",
-//   dom_id: "#swagger-ui",
-//   presets: [
-//     SwaggerUIBundle.presets.apis,
-//     SwaggerUIBundle.SwaggerUIStandalonePreset,
-//   ],
-//   layout: "StandaloneLayout",
-// });
+import {
+  swagger_html,
+  main_js,
+  swagger_ui_bundle_js,
+  swagger_ui_standalone_preset_js,
+} from "./file_cache.ts";
+
+import { swagger_json_cache } from "./swagger_json.ts";
 
 const doc = {
   info: {
@@ -39,7 +31,7 @@ const doc = {
       url: "https://opensource.org/licenses/MIT",
     },
   },
-  host: `localhost:${Server.PORT}`,
+  host: Swagger.SWAGGER_URL,
   basePath: "/",
   schemes: Swagger.APP_SCHEMES,
   consumes: ["application/json"],
@@ -66,26 +58,37 @@ const doc = {
   components: {}, // by default: empty object (OpenAPI 3.x)
 };
 
-export const route = async (app: Application) => {
-  const endpointsFiles = [...Deno.readDirSync(modules_dir)]
-    .filter((dirEntry) => !dirEntry.name.startsWith("#") && !dirEntry.isFile)
-    .map(
-      (dirEntry) =>
-        `${modules_dir}/${dirEntry.name}/${dirEntry.name}.router.ts`,
-    )
-    .filter((dir_path) => {
-      try {
-        return Deno.statSync(dir_path).isFile;
-      } catch (error) {
-        return false;
-      }
-    });
-  await SwagInit(swagger_json_file, endpointsFiles, doc);
-  const swagger_text = await Deno.readTextFile(swagger_json_file);
-  const swagger_json = JSON.parse(swagger_text);
-  app.get("/swagger.json", (_, res) => res.json(swagger_json));
+export const createSwaggerJson = async () => {
+  const result: any = await SwagInit(
+    swagger_json_file,
+    [`${modules_dir}/index.ts`],
+    doc
+  );
+  if (!result.success) {
+    return false;
+  }
   await Deno.remove(swagger_json_file, { recursive: true });
+  return result.data;
+};
 
-  const swagger_html = await Deno.readTextFile(swagger_html_file);
+export const route = async (app: Application) => {
+  // TODO: bagaimana cara membuat swagger.json tanpa harus link url seperti ini, langsung dari variable
+  let swagger_json = {};
+  swagger_json = await createSwaggerJson();
+  if (!swagger_json) {
+    console.log("✅ Use Swagger on Cache!");
+    swagger_json = swagger_json_cache;
+  }
+  app.get("/swagger.json", (_, res) => res.json(swagger_json));
+
   app.get("/swagger", (_, res) => res.send(swagger_html));
+  app.get("/cache/:file", (req, res) => {
+    const { file } = req.params;
+    res.set("content-type", "text/javascript");
+    if (file == "main.js") return res.send(main_js);
+    if (file == "swagger-ui-bundle.js") return res.send(swagger_ui_bundle_js);
+    if (file == "swagger-ui-standalone-preset.js")
+      return res.send(swagger_ui_standalone_preset_js);
+    return res.status(StatusCodes.NOT_FOUND).send("not found...");
+  });
 };
